requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.2
django-autoslug==1.9.8
django-crispy-forms==2.0
django-environ==0.10.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
gunicorn==20.1.0
phonenumbers==8.13.7
Pillow==9.4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
